+++ index.ts
@@ -1,3 +1,4 @@
 export { isPrime } from "./src/is_prime.js";
 export { integerSqrt } from "./src/integer_sqrt.js";
 export { gcd } from "./src/gcd.js";
+export { combination } from "./src/combination.js";